Break out registry into its own module
